requirements.txt
asgiref==3.4.1
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
PyJWT==2.3.0
pyparsing==3.0.6
pytz==2021.3
requests==2.27.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
